+++ rules/ts-only.js
@@ -42,22 +42,6 @@ module.exports = {
         // https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/explicit-module-boundary-types.md
         '@typescript-eslint/explicit-module-boundary-types': 'error',
 
-        // Disallow calling any variable that is typed as any
-        // https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-unsafe-assignment.md
-        '@typescript-eslint/no-unsafe-assignment': 'error',
-
-        // Disallow member access on any variable that is typed as any
-        // https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-unsafe-member-access.md
-        '@typescript-eslint/no-unsafe-member-access': 'error',
-
-        // Disallow returning any from a function
-        // https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-unsafe-return.md
-        '@typescript-eslint/no-unsafe-return': 'error',
-
-        // Disallows calling an any type value
-        // https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-unsafe-call.md
-        '@typescript-eslint/no-unsafe-call': 'error',
-
         // Ensure exhaustiveness checking in switch with union type
         // https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/switch-exhaustiveness-check.md
         '@typescript-eslint/switch-exhaustiveness-check': 'error'