+++ strict.js
@@ -1,0 +1,4 @@
+module.exports = {
+  extends: ['./index', './rules/no-unsafe'].map(require.resolve),
+  rules: {}
+};